backend/requirements.txt
fastapi
uvicorn
python-multipart
python-doctr[torch]
opencv-python-headless
pandas
openpyxl
pdf2image
schwifty
python-stdnum
